+++ src/main/java/commands/AddCommand.java
@@ -1,3 +1,14 @@
+package commands;
+
+import exceptions.InputException;
+import storage.Storage;
+import storage.TaskList;
+import tasks.Deadline;
+import tasks.Event;
+import tasks.Task;
+import tasks.ToDo;
+import ui.Ui;
+
 import java.io.IOException;
 import java.time.LocalDateTime;
