+++ src/main/java/commands/Command.java
@@ -1,3 +1,10 @@
+package commands;
+
+import exceptions.InputException;
+import storage.Storage;
+import storage.TaskList;
+import ui.Ui;
+
 import java.io.IOException;
 
 public interface Command {
