+++ src/main/java/commands/ExitCommand.java
@@ -1,3 +1,9 @@
+package commands;
+
+import storage.Storage;
+import storage.TaskList;
+import ui.Ui;
+
 public class ExitCommand implements Command {
 
     @Override
