+++ src/main/java/commands/HelpCommand.java
@@ -1,3 +1,9 @@
+package commands;
+
+import storage.Storage;
+import storage.TaskList;
+import ui.Ui;
+
 public class HelpCommand implements Command {
 
     @Override
@@ -5,8 +11,8 @@ public class HelpCommand implements Command {
         ui.printLine();
         ui.show("PandaBot Commands:");
         ui.show("1. todo <description> : Adds a new Todo task.");
-        ui.show("2. deadline <description> /by <DD/MM/YYYY HHmm> : Adds a new Deadline task.");
-        ui.show("3. event <task description> /from <DD/MM/YYYY HHmm> /to <DD/MM/YYYY HHmm> : Adds a new Event task.");
+        ui.show("2. deadline <description> /by <DD/MM/YYYY HHmm> : Adds a new tasks.Deadline task.");
+        ui.show("3. event <task description> /from <DD/MM/YYYY HHmm> /to <DD/MM/YYYY HHmm> : Adds a new tasks.Event task.");
         ui.show("4. list : Lists all tasks.");
         ui.show("5. mark <task number> : Marks the specified task as done.");
         ui.show("6. unmark <task number> : Unmarks the specified task.");
