+++ src/main/java/parser/Parser.java
@@ -1,3 +1,8 @@
+package parser;
+
+import commands.*;
+import exceptions.InputException;
+
 public class Parser {
     public static Command parse(String fullCommand) throws InputException {
         String[] commandParts = fullCommand.trim().split(" ", 2);
